# oresight-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OreSight API",
    "description": "Backend API for OreSight - a mine reserve intelligence prototype built for SIH26009 (MOIL manganese mining).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/equipment": {
      "get": {
        "tags": [
          "equipment"
        ],
        "summary": "List equipment",
        "description": "Return equipment, optionally filtered to a single site.",
        "operationId": "list_equipment_equipment_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one site's equipment",
              "title": "Site Id"
            },
            "description": "Filter to one site's equipment"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EquipmentOut"
                  },
                  "title": "Response List Equipment Equipment Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/equipment/{equipment_id}/status": {
      "post": {
        "tags": [
          "equipment"
        ],
        "summary": "Update an equipment's status",
        "description": "Update an equipment's status and reason. Marking it 'down' also opens a\nhigh-severity `equipment_failure` risk event for its site.",
        "operationId": "update_equipment_status_equipment__equipment_id__status_post",
        "parameters": [
          {
            "name": "equipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Equipment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production": {
      "get": {
        "tags": [
          "production"
        ],
        "summary": "List production records",
        "description": "Return production records from the last `days` days, ordered oldest to\nnewest.",
        "operationId": "list_production_records_production_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one site",
              "title": "Site Id"
            },
            "description": "Filter to one site"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "How many trailing days to include",
              "default": 30,
              "title": "Days"
            },
            "description": "How many trailing days to include"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductionRecordOut"
                  },
                  "title": "Response List Production Records Production Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "production"
        ],
        "summary": "Record a day's production",
        "description": "Insert a new production record. Returns 409 if one already exists for\nthis site and date.",
        "operationId": "create_production_record_production_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductionRecordCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionRecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/risk-events": {
      "get": {
        "tags": [
          "risk-events"
        ],
        "summary": "List risk events",
        "description": "Return risk events, newest first, optionally filtered by site and/or\nresolved status.",
        "operationId": "list_risk_events_risk_events_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one site",
              "title": "Site Id"
            },
            "description": "Filter to one site"
          },
          {
            "name": "resolved",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by resolved status",
              "title": "Resolved"
            },
            "description": "Filter by resolved status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiskEventOut"
                  },
                  "title": "Response List Risk Events Risk Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/risk-events/{risk_event_id}/causal-graph": {
      "get": {
        "tags": [
          "risk-events"
        ],
        "summary": "Get the causal graph behind a risk event",
        "description": "Return the causal chain that led to this risk event.\n\nSTUB: replaced on Day 3 with a live Neo4j traversal. For now returns a\nfixed, realistically-shaped 5-node chain so the frontend can build the\nReact Flow graph view today.",
        "operationId": "get_causal_graph_risk_events__risk_event_id__causal_graph_get",
        "parameters": [
          {
            "name": "risk_event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Risk Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CausalGraphOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CausalGraphOut": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "title": "CausalGraphOut",
        "example": {
          "edges": [
            {
              "relationship": "DELAYS",
              "source": "weather_2026_03_14",
              "target": "blast_plan_b204"
            },
            {
              "relationship": "AFFECTS",
              "source": "blast_plan_b204",
              "target": "zone_b2"
            },
            {
              "relationship": "OPERATES_IN",
              "source": "equipment_ex201",
              "target": "zone_b2"
            },
            {
              "relationship": "CAUSES",
              "source": "zone_b2",
              "target": "risk_event_shortfall"
            }
          ],
          "nodes": [
            {
              "id": "weather_2026_03_14",
              "label": "Heavy Rainfall 14 Mar",
              "type": "WeatherEvent"
            },
            {
              "id": "blast_plan_b204",
              "label": "Blast Plan B-204",
              "type": "BlastPlan"
            },
            {
              "id": "zone_b2",
              "label": "Zone B2",
              "type": "OreZone"
            },
            {
              "id": "equipment_ex201",
              "label": "Excavator EX-201",
              "type": "Equipment"
            },
            {
              "id": "risk_event_shortfall",
              "label": "Production shortfall risk",
              "type": "RiskEvent"
            }
          ]
        }
      },
      "EquipmentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "site_id": {
            "type": "integer",
            "title": "Site Id"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "equipment_type": {
            "type": "string",
            "title": "Equipment Type"
          },
          "status": {
            "type": "string",
            "enum": [
              "up",
              "down"
            ],
            "title": "Status"
          },
          "last_status_change": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Status Change"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "site_id",
          "site_name",
          "name",
          "equipment_type",
          "status",
          "last_status_change",
          "status_reason"
        ],
        "title": "EquipmentOut",
        "example": {
          "equipment_type": "excavator",
          "id": 1,
          "last_status_change": "2026-08-30T05:11:27.578760+00:00",
          "name": "Excavator EX-201",
          "site_id": 1,
          "site_name": "Balaghat",
          "status": "down",
          "status_reason": "Hydraulic pump failure - spare part on order, ETA 3 days"
        }
      },
      "EquipmentStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "up",
              "down"
            ],
            "title": "Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "EquipmentStatusUpdate",
        "example": {
          "reason": "Engine overheating - pulled for inspection",
          "status": "down"
        }
      },
      "GraphEdge": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "relationship": {
            "type": "string",
            "title": "Relationship"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "relationship"
        ],
        "title": "GraphEdge"
      },
      "GraphNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "GraphNode"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProductionRecordCreate": {
        "properties": {
          "site_id": {
            "type": "integer",
            "title": "Site Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "actual_output": {
            "type": "number",
            "title": "Actual Output"
          },
          "target_output": {
            "type": "number",
            "title": "Target Output"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "date",
          "actual_output",
          "target_output"
        ],
        "title": "ProductionRecordCreate",
        "example": {
          "actual_output": 1180.0,
          "date": "2026-08-31",
          "site_id": 1,
          "target_output": 1250.0
        }
      },
      "ProductionRecordOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "site_id": {
            "type": "integer",
            "title": "Site Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "actual_output": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Output"
          },
          "target_output": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Output"
          },
          "variance_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variance Pct"
          }
        },
        "type": "object",
        "required": [
          "id",
          "site_id",
          "date",
          "actual_output",
          "target_output",
          "variance_pct"
        ],
        "title": "ProductionRecordOut",
        "example": {
          "actual_output": 712.4,
          "date": "2026-08-29",
          "id": 42,
          "site_id": 1,
          "target_output": 1062.5,
          "variance_pct": -32.96
        }
      },
...
      "RiskEventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "site_id": {
            "type": "integer",
            "title": "Site Id"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "risk_type": {
            "type": "string",
            "title": "Risk Type"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Severity"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "resolved": {
            "type": "boolean",
            "title": "Resolved"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "site_id",
          "site_name",
          "risk_type",
          "severity",
          "score",
          "description",
          "resolved",
          "detected_at"
        ],
        "title": "RiskEventOut",
        "example": {
          "description": "Haul Truck HT-302 engine overheating has taken the unit fully offline.",
          "detected_at": "2026-08-31T05:11:27.578760+00:00",
          "id": 5,
          "resolved": false,
          "risk_type": "equipment_failure",
          "score": 0.93,
          "severity": "critical",
          "site_id": 2,
          "site_name": "Nagpur"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}